+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Pin `mujoco` and `mujoco-warp` dependencies to `~=3.6.0`
 - Update default environment map texture in GL viewer (source: https://polyhaven.com/a/brown_photostudio_02)
 - Increase conveyor rail roughness in `example_basic_conveyor` to reduce mirror-like reflections
+- Migrate all raycast logic to `geometry.raycast`, all raycast functions now return distance and normal information
 
 ### Fixed